requirements.txt
mmcv-full==1.5.2
mmdet==2.24.1
numpy==1.22.4
opencv-python-headless==4.5.5.64
openmim==0.1.5
torch==1.11.0
torchvision==0.12.0
